# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"BearerAuth":[]}],"tags":[{"name":"User management","description":"Endpoints for managing users."},{"name":"Accommodation management","description":"Endpoints for managing accommodations."},{"name":"Booking management","description":"Endpoints for managing booking."}],"paths":{"/users/{id}/role":{"put":{"tags":["User management"],"summary":"Updates the role of a specific user.","operationId":"updateUserRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleUpdateRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/bookings/{id}":{"get":{"tags":["Booking management"],"summary":"Provides information about a specific booking.","operationId":"getBookingById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingResponseDto"}}}}}},"put":{"tags":["Booking management"],"summary":"Allows customers to update booking details beyond status.","operationId":"updateBookingDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingResponseDto"}}}}}},"delete":{"tags":["Booking management"],"summary":"Enables the cancellation of bookings.","operationId":"cancelBooking","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["Booking management"],"summary":"Allows managers to update booking status alone.","operationId":"updateBookingStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingStatusPatchRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingResponseDto"}}}}}}},"/accommodations/{id}":{"get":{"tags":["Accommodation management"],"summary":"Retrieves detailed information about a specific accommodation.","operationId":"getAccommodationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccommodationResponseDto"}}}}}},"put":{"tags":["Accommodation management"],"summary":"Updates all details of a specific accommodation.","operationId":"updateAccommodation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccommodationRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccommodationResponseDto"}}}}}},"delete":{"tags":["Accommodation management"],"summary":"Enables the removal of accommodations by marking them as deleted.","operationId":"deleteAccommodation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/payments":{"get":{"tags":["payment-controller"],"operationId":"getPayments","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentResponseDto"}}}}}}},"post":{"tags":["payment-controller"],"operationId":"initiatePayment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponseDto"}}}}}}},"/payments/renew/{paymentId}":{"post":{"tags":["payment-controller"],"operationId":"renewPaymentSession","parameters":[{"name":"paymentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponseDto"}}}}}}},"/bookings":{"get":{"tags":["Booking management"],"summary":"Retrieves bookings based on user ID and their status for managers.","operationId":"getBookings","parameters":[{"name":"userId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookingResponseDto"}}}}}}},"post":{"tags":["Booking management"],"summary":"Permits the creation of new accommodation bookings.","operationId":"createBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequestDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingResponseDto"}}}}}}},"/auth/register":{"post":{"tags":["User management"],"summary":"Register a new user.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationRequestDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/auth/login":{"post":{"tags":["User management"],"summary":"Login an authenticated user.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserLoginResponseDto"}}}}}}},"/accommodations":{"get":{"tags":["Accommodation management"],"summary":"Provides a list of available accommodations.","operationId":"getAllAccommodations","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccommodationResponseDto"}}}}}}},"post":{"tags":["Accommodation management"],"summary":"Permits the addition of new accommodations.","operationId":"addAccommodation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccommodationRequestDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccommodationResponseDto"}}}}}}},"/users/me/profile":{"patch":{"tags":["User management"],"summary":"Allows users to update their profile information.","operationId":"updateUserProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/users/me/password":{"patch":{"tags":["User management"],"summary":"Allows users to update their password.","operationId":"updateUserPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordUpdateRequestDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/users/me":{"get":{"tags":["User management"],"summary":"Retrieves the profile information for the currently logged-in user.","operationId":"getCurrentUserProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/users/confirm-password":{"get":{"tags":["User management"],"summary":"Confirms password update via email link","operationId":"confirmPasswordUpdate","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/payments/success":{"get":{"tags":["payment-controller"],"operationId":"handlePaymentSuccess","parameters":[{"name":"session_id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/payments/cancel":{"get":{"tags":["payment-controller"],"operationId":"handlePaymentCancel","parameters":[{"name":"session_id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bookings/my":{"get":{"tags":["Booking management"],"summary":"Retrieves bookings for the currently logged-in user.","operationId":"getUserBookings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookingResponseDto"}}}}}}}}},"components":{"schemas":{"UserRoleUpdateRequestDto":{"type":"object","properties":{"role":{"type":"string"}},"required":["role"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}}},"BookingRequestDto":{"type":"object","properties":{"accommodationId":{"type":"integer","format":"int64"},"checkInDate":{"type":"string","format":"date"},"checkOutDate":{"type":"string","format":"date"}},"required":["accommodationId","checkInDate","checkOutDate"]},"BookingResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accommodationId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"checkInDate":{"type":"string","format":"date"},"checkOutDate":{"type":"string","format":"date"},"status":{"type":"string"}}},"AccommodationRequestDto":{"type":"object","properties":{"type":{"type":"string"},"location":{"type":"string"},"size":{"type":"string"},"amenities":{"type":"array","items":{"type":"string"}},"dailyRate":{"type":"number"},"availability":{"type":"integer","format":"int32"}},"required":["availability","dailyRate","location","type"]},"AccommodationResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"location":{"type":"string"},"size":{"type":"string"},"amenities":{"type":"array","items":{"type":"string"}},"dailyRate":{"type":"number"},"availability":{"type":"integer","format":"int32"}}},"PaymentRequestDto":{"type":"object","properties":{"bookingId":{"type":"integer","format":"int64"}},"required":["bookingId"]},"PaymentResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"bookingId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["PENDING","PAID","EXPIRED"]},"sessionUrl":{"type":"string"},"sessionId":{"type":"string"},"amountToPay":{"type":"number"}}},"UserRegistrationRequestDto":{"type":"object","description":"Request DTO for registering a new user.","properties":{"email":{"type":"string","description":"The user's email address.","example":"john.doe@email.com"},"password":{"type":"string","description":"The user's password. It must be between 8 and 20 characters.","example":"SecurePassword123","maxLength":20,"minLength":8},"confirmPassword":{"type":"string","description":"The confirmation password. It should match the password field.","example":"SecurePassword123","maxLength":20,"minLength":8},"firstName":{"type":"string","description":"The first name of the registering user.","example":"John"},"lastName":{"type":"string","description":"The last name of the registering user.","example":"Doe"}},"required":["confirmPassword","email","firstName","lastName","password"]},"UserLoginRequestDto":{"type":"object","description":"Request DTO for a registered user signing in.","properties":{"email":{"type":"string","description":"The user's email address. This must be a valid email format (e.g., example@email.com).","example":"emily@email.com"},"password":{"type":"string","description":"The user's password. It must be between 8 and 20 characters.","example":"PaSsWorD789","maxLength":20,"minLength":8}},"required":["email","password"]},"UserLoginResponseDto":{"type":"object","properties":{"token":{"type":"string"}}},"UserUpdateRequestDto":{"type":"object","properties":{"firstName":{"type":"string","maxLength":50,"minLength":0},"lastName":{"type":"string","maxLength":50,"minLength":0}}},"PasswordUpdateRequestDto":{"type":"object","properties":{"newPassword":{"type":"string","description":"The user's password. It must be between 8 and 20 characters.","example":"SecurePassword123","maxLength":20,"minLength":8}},"required":["newPassword"]},"BookingStatusPatchRequestDto":{"type":"object","properties":{"status":{"type":"string"}},"required":["status"]},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}}},"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}