# src/main/java/mate/academy/accommodationbookingservice/controller/AuthenticationController.java
package mate.academy.accommodationbookingservice.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import mate.academy.accommodationbookingservice.dto.user.request.UserLoginRequestDto;
import mate.academy.accommodationbookingservice.dto.user.request.UserRegistrationRequestDto;
import mate.academy.accommodationbookingservice.dto.user.response.UserLoginResponseDto;
import mate.academy.accommodationbookingservice.dto.user.response.UserResponseDto;
import mate.academy.accommodationbookingservice.exceptions.RegistrationException;
import mate.academy.accommodationbookingservice.security.AuthenticationService;
import mate.academy.accommodationbookingservice.service.user.UserService;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@Tag(name = "User management", description = "Endpoints for managing users.")
@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
public class AuthenticationController {
    private final UserService userService;
    private final AuthenticationService authenticationService;

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    @Operation(summary = "Register a new user.")
    public UserResponseDto register(@RequestBody @Valid UserRegistrationRequestDto request)
            throws RegistrationException {
        return userService.register(request);
    }

    @PostMapping("/login")
    @ResponseStatus(HttpStatus.OK)
    @Operation(summary = "Login an authenticated user.")
    public UserLoginResponseDto login(@RequestBody @Valid UserLoginRequestDto request) {
        return authenticationService.authenticate(request);
    }
}