# src/main/java/mate/academy/accommodationbookingservice/dto/booking/BookingResponseDto.java
package mate.academy.accommodationbookingservice.dto.booking;

import java.time.LocalDate;
import lombok.Data;

@Data
public class BookingResponseDto {
    private Long id;
    private Long accommodationId;
    private Long userId;
    private LocalDate checkInDate;
    private LocalDate checkOutDate;
    private String status;
}