# src/main/java/mate/academy/accommodationbookingservice/dto/payment/PaymentRequestDto.java
package mate.academy.accommodationbookingservice.dto.payment;

import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class PaymentRequestDto {
    @NotNull
    private Long bookingId;
}