# src/main/java/mate/academy/accommodationbookingservice/dto/user/response/UserLoginResponseDto.java
package mate.academy.accommodationbookingservice.dto.user.response;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class UserLoginResponseDto {
    private String token;
}